@@ -10,8 +10,8 @@ class _AnimationsFromScratchScreenState
     extends State<AnimationsFromScratchScreen>
     with SingleTickerProviderStateMixin {
   bool select = true;
-  AnimationController _controller;
-  Animation<Size> _heightAnimation;
+  late AnimationController _controller;
+  late Animation<Size> _heightAnimation;
 
   @override
   void initState() {
