@@ -1,4 +1,4 @@
-package sa.takwen.flutter_animation
+package com.example.flutter_animation
 
 import io.flutter.embedding.android.FlutterActivity
 
