@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_animation/animation_screen/animated_cross_fade_screen.dart';
+import 'package:flutter_animation/animation_screen/animated_dynamic_list_screen.dart';
 import 'package:flutter_animation/animation_screen/fade_animation_screen.dart';
 import 'animation_screen/animated_builder_screen.dart';
 import 'animation_screen/animated_container.dart';
@@ -9,7 +10,6 @@ import 'animation_screen/pageRouteBuilder_screen.dart';
 import 'animation_screen/slide_transition.dart';
 import 'animation_screen/transition_opacity_screen.dart';
 import 'shared/resources/navigation.dart';
-
 
 void main() {
   runApp(MyApp());
@@ -25,60 +25,63 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: MyHomePage(),
+      home: const MyHomePage(),
     );
   }
 }
 
 class MyHomePage extends StatelessWidget {
+  const MyHomePage({Key? key}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       body: CustomScrollView(
         slivers: [
-          SliverAppBar(
+          const SliverAppBar(
             expandedHeight: 200,
             pinned: true,
             flexibleSpace: FlexibleSpaceBar(
               title: Text("Flutter Animation"),
               centerTitle: true,
-              background: FlutterLogo(size:20,),
+              background: FlutterLogo(
+                size: 20,
+              ),
             ),
           ),
           SliverList(
             delegate: SliverChildListDelegate([
               Card(
                 child: ListTile(
-                  title: Text("Set up a PageRouteBuilder"),
+                  title: const Text("Set up a PageRouteBuilder"),
                   onTap: () {
-                    Navigator.of(context).push(createRoute( PageRouteBuilderScreen()));
+                    Navigator.of(context)
+                        .push(createRoute(PageRouteBuilderScreen()));
                   },
                 ),
               ),
               Card(
                 child: GestureDetector(
-                  child: Hero(
+                  child: const Hero(
                     tag: 'imageHero',
                     child: FadeInImage(
                       width: double.infinity,
                       height: 200,
                       fit: BoxFit.cover,
                       placeholder: AssetImage('assets/images/image2.png'),
-                      image: NetworkImage("https://docs.flutter.dev/assets/images/flutter-logo-sharing.png"),
+                      image: NetworkImage(
+                          "https://docs.flutter.dev/assets/images/flutter-logo-sharing.png"),
                     ),
                   ),
                   onTap: () {
                     Navigator.of(context).push(
                       PageRouteBuilder(
-                        transitionDuration: Duration(milliseconds: 1000),
-                        pageBuilder: (
-                            BuildContext context,
+                        transitionDuration: const Duration(milliseconds: 1000),
+                        pageBuilder: (BuildContext context,
                             Animation<double> animation,
                             Animation<double> secondaryAnimation) {
                           return HeroAnimationScreen();
                         },
-                        transitionsBuilder: (
-                            BuildContext context,
+                        transitionsBuilder: (BuildContext context,
                             Animation<double> animation,
                             Animation<double> secondaryAnimation,
                             Widget child) {
@@ -96,61 +99,90 @@ class MyHomePage extends StatelessWidget {
               ),
               Card(
                 child: ListTile(
-                  title: Text("Container Animation Controiler"),
+                  title: const Text("Container Animation Controiler"),
                   onTap: () {
-                    Navigator.push(context,MaterialPageRoute(builder:(context)=>AnimationsFromScratchScreen()));
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) =>
+                                AnimationsFromScratchScreen()));
                   },
                 ),
               ),
               Card(
                 child: ListTile(
-                  title: Text("Using the AnimatedBuilder Widget"),
+                  title: const Text("Using the AnimatedBuilder Widget"),
                   onTap: () {
-                    Navigator.push(context,MaterialPageRoute(builder:(context)=>AnimatedBuilderScreen()));
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => AnimatedBuilderScreen()));
                   },
                 ),
               ),
               Card(
                 child: ListTile(
-                  title: Text("AnimatedContainer"),
+                  title: const Text("AnimatedContainer"),
                   onTap: () {
-                    Navigator.push(context,MaterialPageRoute(builder:(context)=>AnimatedContainerScreen()));
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => AnimatedContainerScreen()));
                   },
                 ),
               ),
               Card(
                 child: ListTile(
-                  title: Text("Animation & Transition Opacity"),
+                  title: const Text("Animation & Transition Opacity"),
                   onTap: () {
-                    Navigator.push(context,MaterialPageRoute(builder:(context)=>TransitionOpacityScreen()));
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => TransitionOpacityScreen()));
                   },
                 ),
               ),
               Card(
                 child: ListTile(
-                  title: Text("Slide Transition"),
+                  title: const Text("Slide Transition"),
                   onTap: () {
-                    Navigator.push(context,MaterialPageRoute(builder:(context)=>SlideTransitionScreen()));
-                  },
-                ),
-              ),
-               Card(
-                child: ListTile(
-                  title: Text("Fade Animation"),
-                  onTap: () {
-                    Navigator.of(context).push(createRoute(FadeAnimationScreen()));
+                    Navigator.push(
+                        context,
+                        MaterialPageRoute(
+                            builder: (context) => SlideTransitionScreen()));
                   },
                 ),
               ),
               Card(
                 child: ListTile(
-                  title: Text("Animated Cross Fade"),
+                  title: const Text("Fade Animation"),
                   onTap: () {
-                    Navigator.of(context).push(createRoute(AnimatedCrossFadeScreen()));
+                    Navigator.of(context)
+                        .push(createRoute(FadeAnimationScreen()));
                   },
                 ),
               ),
-              SizedBox(height: 200,)
+              Card(
+                child: ListTile(
+                  title: const Text("Animated Cross Fade"),
+                  onTap: () {
+                    Navigator.of(context)
+                        .push(createRoute(const AnimatedCrossFadeScreen()));
+                  },
+                ),
+              ),
+              Card(
+                child: ListTile(
+                  title: const Text("Animated Dynamic List"),
+                  onTap: () {
+                    Navigator.of(context)
+                        .push(createRoute(const AnimatedDynamicListScreen()));
+                  },
+                ),
+              ),
+              const SizedBox(
+                height: 200,
+              )
             ]),
           )
         ],
@@ -158,5 +190,3 @@ class MyHomePage extends StatelessWidget {
     );
   }
 }
-
-
